--- a/test/test_ops.py
+++ b/test/test_ops.py
@@ -5,6 +5,7 @@ import unittest
 import numpy as np
 
 import torch
+from functools import lru_cache
 from torch import Tensor
 from torch.autograd import gradcheck
 from torch.nn.modules.utils import _pair
@@ -496,6 +497,7 @@ class DeformConvTester(OpTester, unittest.TestCase):
         out += bias.view(1, n_out_channels, 1, 1)
         return out
 
+    @lru_cache(maxsize=None)
     def get_fn_args(self, device, contiguous, batch_sz, dtype):
         n_in_channels = 6
         n_out_channels = 2